% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ppi.R
\name{read.ppi}
\alias{read.ppi}
\title{Title}
\usage{
read.ppi(ppi_location)
}
\arguments{
\item{ppi_location}{sfsaf}
}
\value{
sdfsdf
}
\description{
Title
}
\examples{
ppi_location<-system.file("extdata", "ppi.csv", package = "my.forth");x <- read.ppi(ppi_location)
}
